move parallization helper to utils
